refactor: SeatPickerPresenter 에서 database 의존성 제거
