+++ app/src/main/java/woowacourse/movie/data/reservation/ReservationRepository.kt
@@ -1,0 +1,11 @@
+package woowacourse.movie.data.reservation
+
+import woowacourse.movie.ui.model.MovieTicketModel
+
+class ReservationRepository(private val localDataSource: ReservationLocalDataSource) {
+    fun insertReservation(ticket: MovieTicketModel) {
+        localDataSource.insertReservation(ticket)
+    }
+
+    fun getReservations(): List<MovieTicketModel> = localDataSource.getReservations()
+}
